+++ src/main/java/com/lucendar/common/types/ByteBufferBackedOutputStream.java
@@ -3,9 +3,16 @@ package com.lucendar.common.types;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
 
+/**
+ * 后端为 ByteBuffer 的 OutputStream
+ */
 public class ByteBufferBackedOutputStream extends OutputStream {
     private final ByteBuffer buf;
 
+    /**
+     * 构造函数
+     * @param buf 作为后端的 ByteBuffer
+     */
     public ByteBufferBackedOutputStream(ByteBuffer buf) {
         this.buf = buf;
     }